Update for changes on Cyclone DDS security branch
The security specification has a mode in which keyhashes containing the
MD5 of the key value are required on the wire, and this requires some
small changes to the topic interface.  Currently, these changes reside
only on the security branch of Cyclone DDS.

The changes use conditional compilation to handle both the master branch
of Cyclone and the security branch to minimise version dependencies.

Signed-off-by: Erik Boasson <eb@ilities.com>