Make sure to reset the error when a typesupport can't be found.
The rcutils error_string is meant to be used like errno; it
may or may not be set when entering a function, calling other
functions may or may not set it, and it is up to the caller
on how to handle all of that.  make_message_value_type() and
make_request_response_value_types() were not resetting the
error properly, and so could leave a failure state around.
Fix that here by mimicing code that exists elsewhere in
rmw_cyclonedds_cpp.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>