Fix use-after-free in error handling bug
The reference to the sertype is not counted and shouldn't be released.
At this point, and in this test anyway, the memory has been freed just
before (by `dds_delete(topic)`) and this probably explains why it hasn't
caused any breakage before.  The assert that failed now is new in
Cyclone.

There has been an ancient interface for creating custom types that did
require it and that was used by `rmw_cyclonedds_cpp` in the past.  The
rewrite of the RMW layer to use one DDS participant per context
(commit 7d640d1a38333e27c538d7efb01192e3f92d8b5d) also dropped support
for that ancient interface.  Instead of removing the #if/unref/#endif,
that commit removed the #if/#endif but left the unref call in place.

Signed-off-by: Erik Boasson <eb@ilities.com>