Link to Cyclone DDS in Quality Declaration
The Cyclone DDS quality declaration got merged in the Cyclone DDS main
repository a long time ago, but no-one remembered to remove it from the
RMW directory, leading to inconsistencies, in, e.g., quality levels.

This removes the obsolete copy from the RMW repository and updates the
quality declaration for the RMW layer to point to the one in the main
Cyclone DDS repository.  It also updates the quality level to correspond
to the one in the Cyclone DDS quality declaration.

Signed-off-by: Erik Boasson <eb@ilities.com>